# docs/openapi/reference/missions_api_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Missions API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/missions": {
      "post": {
        "tags": [
          "Missions",
          "Missions_Analysis"
        ],
        "summary": "Create And Analyze Mission",
        "operationId": "create_and_analyze_mission_missions_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_and_analyze_mission_missions_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Missions",
          "Missions"
        ],
        "summary": "List Missions",
        "description": "Liste les missions avec pagination. Supporte un filtre optionnel par statut.",
        "operationId": "list_missions_missions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Status"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Nombre d'enregistrements \u00e0 sauter",
              "default": 0,
              "title": "Skip"
            },
            "description": "Nombre d'enregistrements \u00e0 sauter"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Nombre max de missions retourn\u00e9es",
              "default": 50,
              "title": "Limit"
            },
            "description": "Nombre max de missions retourn\u00e9es"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Missions",
          "Missions"
        ],
        "summary": "Delete All Missions",
        "description": "Supprime toutes les missions et leur historique (r\u00e9serv\u00e9 aux admins).",
        "operationId": "delete_all_missions_missions_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
...
    "/missions/{mission_id}/status": {
      "patch": {
        "tags": [
          "Missions",
          "Missions"
        ],
        "summary": "Update Mission Status",
        "description": "Modifie le statut d'une mission (r\u00e9serv\u00e9 aux r\u00f4les commercial et admin).\nV\u00e9rifie la validit\u00e9 de la transition avant application et enregistre l'audit.",
        "operationId": "update_mission_status_missions__mission_id__status_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Mission Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MissionStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/missions/{mission_id}/status/history": {
      "get": {
        "tags": [
          "Missions",
          "Missions"
        ],
        "summary": "Get Mission Status History",
        "description": "Retourne l'historique complet des changements de statut d'une mission (audit trail).",
        "operationId": "get_mission_status_history_missions__mission_id__status_history_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Mission Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StatusHistoryEntry"
                  },
                  "title": "Response Get Mission Status History Missions  Mission Id  Status History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/missions/{mission_id}": {
      "get": {
        "tags": [
          "Missions",
          "Missions"
        ],
        "summary": "Get Mission",
        "operationId": "get_mission_missions__mission_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Mission Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissionAnalyzeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Missions",
          "Missions"
        ],
        "summary": "Delete Mission",
        "description": "Supprime une mission sp\u00e9cifique et son historique (r\u00e9serv\u00e9 aux admins et commerciaux).",
        "operationId": "delete_mission_missions__mission_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Mission Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_create_and_analyze_mission_missions_post": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "Body_create_and_analyze_mission_missions_post"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MissionAnalyzeResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "extracted_competencies": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Extracted Competencies"
          },
          "prefiltered_candidates": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prefiltered Candidates",
            "default": []
          },
          "proposed_team": {
            "items": {
              "$ref": "#/components/schemas/TeamMember"
            },
            "type": "array",
            "title": "Proposed Team"
          },
          "fallback_full_scan": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Full Scan",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "status",
          "extracted_competencies",
          "proposed_team"
        ],
        "title": "MissionAnalyzeResponse"
      },
      "MissionStatusUpdate": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "MissionStatusUpdate"
      },
      "StatusHistoryEntry": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "mission_id": {
            "type": "integer",
            "title": "Mission Id"
          },
          "old_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Old Status"
          },
          "new_status": {
            "type": "string",
            "title": "New Status"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "changed_by": {
            "type": "string",
            "title": "Changed By"
          },
          "changed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Changed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "mission_id",
          "new_status",
          "changed_by",
          "changed_at"
        ],
        "title": "StatusHistoryEntry"
      },
      "TaskResponse": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "status"
        ],
        "title": "TaskResponse"
      },
      "TeamMember": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "justification": {
            "type": "string",
            "title": "Justification"
          },
          "estimated_days": {
            "type": "integer",
            "title": "Estimated Days"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "role",
          "justification",
          "estimated_days"
        ],
        "title": "TeamMember"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}