# missions_api/src/missions/crud_router.py
import shared.database as database
from fastapi import APIRouter, Depends, HTTPException, Query
from sqlalchemy import delete, func, select
from sqlalchemy.ext.asyncio import AsyncSession
from shared.auth.jwt import verify_jwt

from .models import (ALLOWED_TRANSITIONS, STATUS_UPDATE_ROLES, Mission,
                     MissionStatus, MissionStatusHistory)
from .schemas import (MissionAnalyzeResponse, MissionStatusUpdate,
                      StatusHistoryEntry)
from .user_router import invalidate_user_active_cache

router = APIRouter(prefix="", tags=["Missions"], dependencies=[Depends(verify_jwt)])


@router.get("/missions")
async def list_missions(
    db: AsyncSession = Depends(database.get_db),
    status: str = None,
    skip: int = Query(0, ge=0, description="Nombre d'enregistrements à sauter"),
    limit: int = Query(50, ge=1, le=500, description="Nombre max de missions retournées"),
    _: dict = Depends(verify_jwt),
):
    """Liste les missions avec pagination. Supporte un filtre optionnel par statut."""
    base_query = select(Mission)
    if status:
        base_query = base_query.where(Mission.status == status)

    total = (await db.execute(
        select(func.count()).select_from(base_query.subquery())
    )).scalar_one()

    query = base_query.order_by(Mission.created_at.desc()).offset(skip).limit(limit)
    result = await db.execute(query)
    missions = result.scalars().all()

    missions_data = []
    for m in missions:
        missions_data.append({
            "id": m.id,
            "title": m.title,
            "description": (
                (m.description[:200] + "...")
                if m.description and len(m.description) > 200
                else m.description
            ),
            "status": m.status or MissionStatus.STAFFED,
            "extracted_competencies": m.extracted_competencies or [],
            "prefiltered_candidates": m.prefiltered_candidates or [],
            "proposed_team": m.proposed_team or [],
            "fallback_full_scan": m.fallback_full_scan,
        })
    return {"items": missions_data, "total": total, "skip": skip, "limit": limit}


@router.patch("/missions/{mission_id}/status")
async def update_mission_status(
    mission_id: int,
    payload: MissionStatusUpdate,
    db: AsyncSession = Depends(database.get_db),
    token_payload: dict = Depends(verify_jwt),
):
    """Modifie le statut d'une mission (réservé aux rôles commercial et admin).
    Vérifie la validité de la transition avant application et enregistre l'audit.
    """
    user_role = token_payload.get("role", "user")
    user_email = token_payload.get("sub", "unknown@zenika.com")

    if user_role not in STATUS_UPDATE_ROLES:
        raise HTTPException(
            status_code=403,
            detail=f"Accès refusé : seuls les rôles {STATUS_UPDATE_ROLES} peuvent modifier le statut d'une mission.",
        )

    result = await db.execute(select(Mission).where(Mission.id == mission_id))
    mission = result.scalars().first()
    if not mission:
        raise HTTPException(status_code=404, detail="Mission introuvable")

    current_status = mission.status or MissionStatus.STAFFED
    new_status = payload.status

    # Validate transition
    allowed = ALLOWED_TRANSITIONS.get(current_status, [])
    if new_status not in allowed:
        raise HTTPException(
            status_code=422,
            detail=f"Transition invalide : '{current_status}' → '{new_status}'. Transitions autorisées : {allowed}",
        )

    # Apply status change
    old_status = current_status
    mission.status = new_status

    # Persist audit entry
    history_entry = MissionStatusHistory(
        mission_id=mission_id,
        old_status=old_status,
        new_status=new_status,
        reason=payload.reason,
        changed_by=user_email,
    )
    db.add(history_entry)
    await db.commit()
    await db.refresh(mission)

    # Invalider le cache actif de tous les membres de la mission modifiée
    for member in (mission.proposed_team or []):
        try:
            uid = int(member.get("user_id", -1))
            if uid > 0:
                await invalidate_user_active_cache(uid)
        except (ValueError, TypeError):
            pass

    return {
        "id": mission.id,
        "status": mission.status,
        "old_status": old_status,
        "changed_by": user_email,
        "reason": payload.reason,
    }


@router.get("/missions/{mission_id}/status/history", response_model=list[StatusHistoryEntry])
async def get_mission_status_history(
    mission_id: int,
    db: AsyncSession = Depends(database.get_db),
    token_payload: dict = Depends(verify_jwt),
):
    """Retourne l'historique complet des changements de statut d'une mission (audit trail)."""
    result = await db.execute(select(Mission).where(Mission.id == mission_id))
    mission = result.scalars().first()
    if not mission:
        raise HTTPException(status_code=404, detail="Mission introuvable")

    history_result = await db.execute(
        select(MissionStatusHistory)
        .where(MissionStatusHistory.mission_id == mission_id)
        .order_by(MissionStatusHistory.changed_at.asc())
    )
    return history_result.scalars().all()


@router.get("/missions/{mission_id}", response_model=MissionAnalyzeResponse)
async def get_mission(mission_id: int, db: AsyncSession = Depends(database.get_db), _: dict = Depends(verify_jwt)):
    result = await db.execute(select(Mission).where(Mission.id == mission_id))
    m = result.scalars().first()
    if not m:
        raise HTTPException(status_code=404, detail="Mission introuvable")
    return {
        "id": m.id,
        "title": m.title,
        "description": m.description,
        "status": m.status or MissionStatus.STAFFED,
        "extracted_competencies": m.extracted_competencies or [],
        "prefiltered_candidates": m.prefiltered_candidates or [],
        "proposed_team": m.proposed_team or [],
    }


@router.delete("/missions")
async def delete_all_missions(db: AsyncSession = Depends(database.get_db), token_payload: dict = Depends(verify_jwt)):
    """Supprime toutes les missions et leur historique (réservé aux admins)."""
    user_role = token_payload.get("role", "user")
    if user_role != "admin":
        raise HTTPException(status_code=403, detail="Accès refusé : rôle admin requis.")

    await db.execute(delete(MissionStatusHistory))
    result_missions = await db.execute(select(Mission))
    all_missions = result_missions.scalars().all()
    await db.execute(delete(Mission))
    await db.commit()
    # Invalider le cache de tous les membres affectés
    seen_uids: set[int] = set()
    for m in all_missions:
        for member in (m.proposed_team or []):
            try:
                uid = int(member.get("user_id", -1))
                if uid > 0 and uid not in seen_uids:
                    seen_uids.add(uid)
                    await invalidate_user_active_cache(uid)
            except (ValueError, TypeError):
                pass
    return {"status": "cleared"}


@router.delete("/missions/{mission_id}")
async def delete_mission(
    mission_id: int,
    db: AsyncSession = Depends(database.get_db),
    token_payload: dict = Depends(verify_jwt),
):
    """Supprime une mission spécifique et son historique (réservé aux admins et commerciaux)."""
    user_role = token_payload.get("role", "user")
    if user_role not in ("admin", "commercial"):
        raise HTTPException(status_code=403, detail="Accès refusé : rôle admin ou commercial requis.")

    result = await db.execute(select(Mission).where(Mission.id == mission_id))
    mission = result.scalars().first()
    if not mission:
        raise HTTPException(status_code=404, detail="Mission introuvable")

    await db.execute(delete(MissionStatusHistory).where(MissionStatusHistory.mission_id == mission_id))
    proposed_team = mission.proposed_team or []
    await db.execute(delete(Mission).where(Mission.id == mission_id))
    await db.commit()
    # Invalider le cache des membres de la mission supprimée
    for member in proposed_team:
        try:
            uid = int(member.get("user_id", -1))
            if uid > 0:
                await invalidate_user_active_cache(uid)
        except (ValueError, TypeError):
            pass
    return {"status": "deleted", "id": mission_id}


@router.post("/missions/recover-stuck")
async def recover_stuck_missions(
    db: AsyncSession = Depends(database.get_db),
    token_payload: dict = Depends(verify_jwt),
    stuck_minutes: int = Query(
        10, ge=1, le=1440,
        description="Missions bloquées depuis plus de N minutes (défaut 10 min)",
    ),
):
    """Watchdog : remet en DRAFT les missions zombies bloquées en ANALYSIS_IN_PROGRESS.

    Une mission est considérée zombie si elle est en ANALYSIS_IN_PROGRESS depuis plus de
    `stuck_minutes` minutes (signal d'une analyse Gemini ayant échoué silencieusement
    ou d'un pod Cloud Run tué pendant le background task).

    Réservé aux admins. Audit trail complet.
    """
    from datetime import datetime, timezone, timedelta
    from sqlalchemy import and_

    user_role = token_payload.get("role", "user")
    user_email = token_payload.get("sub", "unknown@zenika.com")
    if user_role != "admin":
        raise HTTPException(status_code=403, detail="Accès refusé : rôle admin requis.")

    cutoff = datetime.now(timezone.utc).replace(tzinfo=None) - timedelta(minutes=stuck_minutes)

    result = await db.execute(
        select(Mission).where(
            and_(
                Mission.status == MissionStatus.ANALYSIS_IN_PROGRESS,
                Mission.created_at <= cutoff,
            )
        )
    )
    stuck_missions = result.scalars().all()

    recovered = []
    for mission in stuck_missions:
        mission.status = MissionStatus.DRAFT
        history_entry = MissionStatusHistory(
            mission_id=mission.id,
            old_status=MissionStatus.ANALYSIS_IN_PROGRESS,
            new_status=MissionStatus.DRAFT,
            reason=(
                f"Auto-recovery watchdog : zombie détecté après {stuck_minutes} min "
                "— analyse Gemini n'a pas pu compléter (pod tué ou erreur silencieuse)."
            ),
            changed_by=user_email,
        )
        db.add(history_entry)
        recovered.append({"id": mission.id, "title": mission.title})

    if recovered:
        await db.commit()

    return {
        "recovered": len(recovered),
        "stuck_minutes_threshold": stuck_minutes,
        "missions": recovered,
    }